requirements.txt
fastapi
uvicorn
sqlmodel
pymysql
alembic
python_dotenv